+++ x-pack/legacy/plugins/apm/public/components/shared/Links/apm/APMLink.tsx
@@ -8,9 +8,9 @@ import { EuiLink, EuiLinkAnchorProps } from '@elastic/eui';
 import React from 'react';
 import url from 'url';
 import { pick } from 'lodash';
-import { useLocation } from '../../../hooks/useLocation';
-import { APMQueryParams, toQuery, fromQuery } from './url_helpers';
-import { TIMEPICKER_DEFAULTS } from '../../../context/UrlParamsContext/constants';
+import { useLocation } from '../../../../hooks/useLocation';
+import { APMQueryParams, toQuery, fromQuery } from '../url_helpers';
+import { TIMEPICKER_DEFAULTS } from '../../../../context/UrlParamsContext/constants';
 
 interface Props extends EuiLinkAnchorProps {
   path?: string;
